Added feature to AsciiIO: if the met or pro files contain data that is newer than the current time step they are deleted and recreated (this will happen when you restart a simulation).
A bug in ImisDBIO has been detected when writing aggregated profile data: if there is a hoar layer but overall there are less than 5 layers (nE) then the array Pdata was accessed with an index out of bounds. A quick fix has been applied so that at least no rubbish will be written, but a more sophisticated solution will have to be put into place.