We introduced a switch named DETECT_GRASS ([SnowpackAdvanced] section, default value false). The switch is automatically set to true in operational mode unless specified otherwise in io.ini.
The average values needed to detect growing grass are now computed using the COPY mechanism of MeteoIO as well as the (left) mean_avg filter.
Accordingly some renaming took place. We also took the opportunity to rename Mdata.hs1 into Mdata.hs.

Finally, boolean keys in SnowpackConfig are now set to either true or false and not longer to 1 and 0