This commit moves the phase change exclusively to PhaseChange.cc, by moving the flag setting to PhaseChange.cc.
A difference now is that the flags were set before the last iteration, making a difference. I will make an issue, because it may be interesting now to check whether it is still necessary to force the element temperatures to melting temperature when water is present. This is the only remnant now present in SnowPack.cc