Instead of a globally valid freezing and melting temperature, this commit enables the possibility of having different values for freezing and melting temperatures per layer (potentially important for phase changes).
- Especially made to enable the possibility for a freezing point depression in soil.
- Two variables are added to the ElementData class (freezing_tk and melting_tk).
- In the constructor of the ElementData class, these variabels are intitilized with Constants::freezing_tk and Constants::melting_tk (so they have the same values as used in SNOWPACK before this commit.)
- One can specify whenever needed somewhere in SNOWPACK that melting_tk and freezing_tk have other values than the defaults for a specific layer.

Note 1: because the values are initialized with the values that were used before, SNOWPACK will produce exactly the same results as before this commit.

Note 2: two functions in Laws_sn.cc still refer to Constants::melting_tk, because they don't have the ElementData class. These are typical snow functions, so it is not really necessary to have it in there, as the intended use of these variables is for soil.