Allowing to specify the lower boundary condition for Richards equation in the ini file, using the key LB_COND_WATERFLUX. Implemented choices are: WATERTABLE (recommended), FREEDRAINAGE (recommended and default), DIRICHLET, GRAVITATIONALDRAINAGE and SEEPAGE.
DIRICHLET behaviour changed a bit. It now takes the values at the boundaries and keep those constant over the time step and doesn't prescribe saturation anymore. WATERTABLE explicitly prescribes a saturated Dirichlet lower boundary condition.

There is the option NEUMANN available to prescribe a constant flux. However, there are not many interesting applications for this, so an advanced user should dive in the source code when he wants to enable this one. However, it is in there for technical reasons, as FREEDRAINAGE, GRAVITATIONALDRAINAGE and SEEPAGE are Neumann-type boundary conditions.