Update CaamlIO from Version 5 to Version 6.0.3. Not downward compatible.
DONE: 
-read and write caaml-files Version 6.0.3 works (Version 5 does not work anymore)

-reading of grain shape codes was extended and improved

-bugfix rotation of z-axis: 
snowpack needs the data bottom to top with 0 at the ground. in the caaml-files the orientation can be top to bottom with 0 at the top. The transformation from "top-to-bottom" to "bottom-to-top" was done just by reversing the order of the data. Fixed.

-bugfix resampling: 
In the caaml-file the sampling can be different for the different profiles (stratProfile, tempProfile,...). Thus the profiles have to be resampled to one reference-sampling. Here the stratProfile was chosen as the reference. (A different reference sampling could be realized, if useful!!!???) The resampling was faulty. Now the resampling was moved to a new class "DataResampler" (Is there a useful place to move this class to???). Different methods for point-data like temperature and range-data like density are realized. 

-following profiles are written out:
   -stratProfile
   -tempProfile
   -densityProfile
   -lwcProfile (new)
   -specSurfAreaProfile (new)
   -strengthProfile (new)
   -not anymore written is the hardnessProfile

-following profiles are read in:
   -stratProfile
   -tempProfile
   -densityProfile
   -lwcProfile if available

-read and write was tested with 3 different files so far. 
-written files were tested with scheme-file validator.


TO DO:
-if available, read custom-snowpack-data (e.g. sphericity, dendricity,marker,CDot) from stratMetaData, otherwise estimate sphericity, dendricity, marker from grainshape
-write and read depositionDate in Custom-snowpack-data: <snp:DepositionDate>. What about <caaml:validFormationTime>? Which is the correct location?
-write dateTimeReport (time when caaml was created): time is always in UTC. Why not correct time zone?
-when reading: Is check prefix "SLF" useful? Or alwys "uom"?
-clean up the code. add comments. remove debugcode.
-reorganize code. separate xml-code. use constants for paths and names which can be used for reading and writing.
-more testing: different caaml-files (where to get them?) check visualization of written caaml with niviz.
-switch to different XML-library (pugixml)
-move the new class DataResampler to a useful location.